refactoring project structure: moved data files
