+++ lib/ewelink-helper.js
@@ -2,7 +2,7 @@ const crypto = require('crypto');
 const CryptoJS = require('crypto-js');
 const random = require('random');
 
-const { APP_SECRET } = require('./constants');
+const { APP_SECRET } = require('../src/data/constants');
 
 const DEVICE_TYPE_UUID = require('./data/devices-type-uuid');
 const DEVICE_CHANNEL_LENGTH = require('./data/devices-channel-length');
