+++ lib/payloads/wssLoginPayload.js
@@ -1,4 +1,4 @@
-const { APP_ID } = require('../constants');
+const { APP_ID } = require('../../src/data/constants');
 const { timestamp, nonce } = require('../helpers');
 
 const wssLoginPayload = ({ at, apiKey }) => {
