+++ src/classes/DevicePowerUsageRaw.js
@@ -1,7 +1,7 @@
 const WebSocket = require('./WebSocket');
 const payloads = require('../../lib/payloads');
 const { _get } = require('../../lib/helpers');
-const errors = require('../../lib/errors');
+const errors = require('../data/errors');
 
 class DevicePowerUsageRaw extends WebSocket {
   /**
