+++ src/mixins/getDeviceChannelCount.js
@@ -1,5 +1,5 @@
 const { _get } = require('../../lib/helpers');
-const errors = require('../../lib/errors');
+const errors = require('../data/errors');
 
 const { getDeviceChannelCount } = require('../../lib/ewelink-helper');
