+++ src/mixins/setDevicePowerState.js
@@ -1,6 +1,6 @@
-const { APP_ID } = require('../../lib/constants');
+const { APP_ID } = require('../data/constants');
 const { _get, timestamp, nonce } = require('../../lib/helpers');
-const errors = require('../../lib/errors');
+const errors = require('../data/errors');
 
 const { getDeviceChannelCount } = require('../../lib/ewelink-helper');
