+++ test/valid-credentials.spec.js
@@ -1,7 +1,7 @@
 const delay = require('delay');
 
 const ewelink = require('../main');
-const errors = require('../lib/errors');
+const errors = require('../src/data/errors');
 
 const {
   email,
